+++ foundation/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "foundation"
+version = "0.1.0"
+authors = ["Mitchell <mitchellhardie1@gmail.com>","michael-bailey <mickyb18a@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+
+[dependencies]
+regex = "1"
+crossbeam = "0.8.0"
+crossbeam-channel = "0.5.0"
+crossbeam-queue = "0.3.1"
+parking_lot = "0.11.1"
+dashmap = "4.0.2"
+rayon = "1.3.1"
+zeroize = "1.1.0"
+crossterm = "0.19.0"
+log = "0.4"
+url = "2.2.0"
+uuid = {version = "0.8", features = ["serde", "v4"]}
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+
