+++ foundation/src/prelude.rs
@@ -1,0 +1,8 @@
+pub trait IMessagable<TMessage, TSender> {
+	fn send_message(&self, msg: TMessage);
+  fn set_sender(&self, sender: TSender);
+}
+
+pub trait ICooperative {
+	fn tick(&self);
+}
