--- src/foundation/mod.rs
@@ -1,7 +1,0 @@
-pub trait IMessagable<M> {
-	fn send_message(&self, msg: M);
-}
-
-pub trait ICooperative {
-	fn tick(&self);
-}
